requirements.txt
wxPython~=4.1.1
python-dateutil~=2.8.1
easyocr~=1.2.1
opencv-python~=4.4.0.46
numpy~=1.19.4
beautifulsoup4~=4.9.3
fpdf~=1.7.2